Getting close to a rework on how read models are updated
