+++ Source/EventFlow/EventFlow.csproj
@@ -121,26 +121,25 @@
     <Compile Include="Queries\IQueryProcessor.cs" />
     <Compile Include="Queries\QueryProcessor.cs" />
     <Compile Include="Queries\ReadModelByIdQuery.cs" />
-    <Compile Include="ReadStores\V2\IInMemoryReadStoreV2.cs" />
-    <Compile Include="ReadStores\V2\InMemoryReadStoreV2.cs" />
+    <Compile Include="ReadStores\IInMemoryReadStore.cs" />
+    <Compile Include="ReadStores\InMemoryReadStore.cs" />
     <Compile Include="ReadStores\InMemory\Queries\InMemoryQuery.cs" />
     <Compile Include="ReadStores\InMemory\InMemoryQueryHandler.cs" />
     <Compile Include="ReadStores\ILocateByAggregateId.cs" />
     <Compile Include="ReadStores\IReadModelDomainEventApplier.cs" />
     <Compile Include="ReadStores\IReadModelPopulator.cs" />
-    <Compile Include="ReadStores\V2\IReadModelStoreV2.cs" />
+    <Compile Include="ReadStores\IReadModelStore.cs" />
+    <Compile Include="ReadStores\IReadStoreManager.cs" />
     <Compile Include="ReadStores\LocateByAggregateId.cs" />
     <Compile Include="ReadStores\IReadModelLocator.cs" />
     <Compile Include="ReadStores\ReadModelDomainEventApplier.cs" />
-    <Compile Include="ReadStores\V2\MultipleAggregateReadStoreManager.cs" />
-    <Compile Include="ReadStores\V2\ReadModelEnvelope.cs" />
+    <Compile Include="ReadStores\MultipleAggregateReadStoreManager.cs" />
+    <Compile Include="ReadStores\ReadModelEnvelope.cs" />
     <Compile Include="ReadStores\ReadModelPopulator.cs" />
     <Compile Include="ReadStores\ReadModelStore.cs" />
-    <Compile Include="ReadStores\V2\ReadModelStoreV2.cs" />
     <Compile Include="ReadStores\ReadModelUpdate.cs" />
-    <Compile Include="ReadStores\SingleAggregateReadModelStore.cs" />
-    <Compile Include="ReadStores\V2\ReadStoreManagerV2.cs" />
-    <Compile Include="ReadStores\V2\SingleAggregateReadStoreManager.cs" />
+    <Compile Include="ReadStores\ReadStoreManager.cs" />
+    <Compile Include="ReadStores\SingleAggregateReadStoreManager.cs" />
     <Compile Include="Subscribers\DispatchToEventSubscribers.cs" />
     <Compile Include="Aggregates\DomainEvent.cs" />
     <Compile Include="EventStores\EventStore.cs" />
@@ -155,12 +154,7 @@
     <Compile Include="MetadataProviders\AddEventTypeMetadataProvider.cs" />
     <Compile Include="MetadataProviders\AddMachineNameMetadataProvider.cs" />
     <Compile Include="MetadataProviders\AddGuidMetadataProvider.cs" />
-    <Compile Include="ReadStores\InMemory\InMemoryReadModelStore.cs" />
-    <Compile Include="ReadStores\InMemory\IInMemoryReadModelStore.cs" />
     <Compile Include="ReadStores\IReadModel.cs" />
-    <Compile Include="ReadStores\IReadStoreManager.cs" />
-    <Compile Include="ReadStores\MultiAggregateReadModelStore.cs" />
-    <Compile Include="ReadStores\ReadStoreManager.cs" />
     <Compile Include="EventStores\SerializedEvent.cs" />
     <Compile Include="EventStores\EventDefinition.cs" />
     <Compile Include="EventStores\EventDefinitionService.cs" />
@@ -181,7 +175,6 @@
     <Compile Include="Subscribers\IDomainEventPublisher.cs" />
     <Compile Include="Subscribers\ISubscribeSynchronousTo.cs" />
     <Compile Include="ReadStores\IReadModelContext.cs" />
-    <Compile Include="ReadStores\IReadModelStore.cs" />
     <Compile Include="EventStores\IUncommittedEvent.cs" />
     <Compile Include="Aggregates\Metadata.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
