+++ Source/EventFlow/Extensions/EventFlowOptionsReadStoresExtensions.cs
@@ -21,8 +21,8 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 using EventFlow.Configuration.Registrations;
+using EventFlow.Queries;
 using EventFlow.ReadStores;
-using EventFlow.ReadStores.V2;
 
 namespace EventFlow.Extensions
 {
@@ -30,24 +30,26 @@ namespace EventFlow.Extensions
     {
         public static EventFlowOptions UseReadStoreFor<TReadStore, TReadModel>(
             this EventFlowOptions eventFlowOptions)
-            where TReadStore : class, IReadModelStoreV2<TReadModel>
+            where TReadStore : class, IReadModelStore<TReadModel>
             where TReadModel : class, IReadModel, new()
         {
             return eventFlowOptions.RegisterServices(f =>
                 {
                     f.Register<IReadStoreManager, SingleAggregateReadStoreManager<TReadStore, TReadModel>>();
+                    f.Register<IQueryHandler<ReadModelByIdQuery<TReadModel>, TReadModel>, ReadModelByIdQueryHandler<TReadStore, TReadModel>>();
                 });
         }
 
         public static EventFlowOptions UseReadStoreFor<TReadStore, TReadModel, TReadModelLocator>(
             this EventFlowOptions eventFlowOptions)
-            where TReadStore : class, IReadModelStoreV2<TReadModel>
+            where TReadStore : class, IReadModelStore<TReadModel>
             where TReadModel : class, IReadModel, new()
             where TReadModelLocator : IReadModelLocator
         {
             return eventFlowOptions.RegisterServices(f =>
                 {
                     f.Register<IReadStoreManager, MultipleAggregateReadStoreManager<TReadStore, TReadModel, TReadModelLocator>>();
+                    f.Register<IQueryHandler<ReadModelByIdQuery<TReadModel>, TReadModel>, ReadModelByIdQueryHandler<TReadStore, TReadModel>>();
                 });
         }
 
@@ -56,8 +58,8 @@ namespace EventFlow.Extensions
             where TReadModel : class, IReadModel, new()
         {
             return eventFlowOptions
-                .RegisterServices(f => f.Register<IInMemoryReadStoreV2<TReadModel>, InMemoryReadStoreV2<TReadModel>>(Lifetime.Singleton))
-                .UseReadStoreFor<IInMemoryReadStoreV2<TReadModel>, TReadModel>();
+                .RegisterServices(f => f.Register<IInMemoryReadStore<TReadModel>, InMemoryReadStore<TReadModel>>(Lifetime.Singleton))
+                .UseReadStoreFor<IInMemoryReadStore<TReadModel>, TReadModel>();
         }
 
         public static EventFlowOptions UseInMemoryReadStoreFor<TReadModel, TReadModelLocator>(
@@ -66,42 +68,8 @@ namespace EventFlow.Extensions
             where TReadModelLocator : IReadModelLocator
         {
             return eventFlowOptions
-                .RegisterServices(f => f.Register<IInMemoryReadStoreV2<TReadModel>, InMemoryReadStoreV2<TReadModel>>(Lifetime.Singleton))
-                .UseReadStoreFor<IInMemoryReadStoreV2<TReadModel>, TReadModel, TReadModelLocator>();
+                .RegisterServices(f => f.Register<IInMemoryReadStore<TReadModel>, InMemoryReadStore<TReadModel>>(Lifetime.Singleton))
+                .UseReadStoreFor<IInMemoryReadStore<TReadModel>, TReadModel, TReadModelLocator>();
         }
-
-
-        /*
-        public static EventFlowOptions UseInMemoryReadStoreFor<TReadModel, TReadModelLocator>(
-            this EventFlowOptions eventFlowOptions)
-            where TReadModel : IReadModel, new()
-            where TReadModelLocator : IReadModelLocator
-        {
-            eventFlowOptions.AddReadModelStore<IInMemoryReadModelStore<TReadModel>>();
-            eventFlowOptions.RegisterServices(f =>
-                {
-                    f.Register<IInMemoryReadModelStore<TReadModel>, InMemoryReadModelStore<TReadModel, TReadModelLocator>>(Lifetime.Singleton);
-                    f.Register<IQueryHandler<InMemoryQuery<TReadModel>, IEnumerable<TReadModel>>, InMemoryQueryHandler<TReadModel>>();
-                    f.Register<IQueryHandler<ReadModelByIdQuery<TReadModel>, TReadModel>, InMemoryQueryHandler<TReadModel>>();
-                });
-            return eventFlowOptions;
-        }
-
-        public static EventFlowOptions AddReadModelStore<TReadModelStore>(
-            this EventFlowOptions eventFlowOptions,
-            Lifetime lifetime = Lifetime.AlwaysUnique)
-            where TReadModelStore : class, IReadModelStore
-        {
-            if (typeof(TReadModelStore).IsInterface)
-            {
-                eventFlowOptions.RegisterServices(f => f.Register<IReadModelStore>(r => r.Resolver.Resolve<TReadModelStore>(), lifetime));
-            }
-            else
-            {
-                eventFlowOptions.RegisterServices(f => f.Register<IReadModelStore, TReadModelStore>(lifetime));
-            }
-
-            return eventFlowOptions;
-        }*/
     }
 }
