+++ Source/EventFlow/ReadStores/IInMemoryReadStore.cs
@@ -20,32 +20,10 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using System;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using EventFlow.Aggregates;
-
-namespace EventFlow.ReadStores.V2
+namespace EventFlow.ReadStores
 {
-    public interface IReadModelStoreV2<TReadModel>
+    public interface IInMemoryReadStore<TReadModel> : IReadModelStore<TReadModel>
         where TReadModel : class, IReadModel, new()
     {
-        Task<ReadModelEnvelope<TReadModel>> GetAsync(
-            string id,
-            CancellationToken cancellationToken);
-
-        Task DeleteAsync(
-            string id,
-            CancellationToken cancellationToken);
-
-        Task DeleteAllAsync(
-            CancellationToken cancellationToken);
-
-        Task UpdateAsync(
-            IReadOnlyCollection<ReadModelUpdate> readModelUpdates,
-            IReadModelContext readModelContext,
-            Func<IReadModelContext, IReadOnlyCollection<IDomainEvent>, ReadModelEnvelope<TReadModel>, CancellationToken, Task<ReadModelEnvelope<TReadModel>>> updateReadModel,
-            CancellationToken cancellationToken);
     }
 }
