+++ Source/EventFlow/ReadStores/ReadModelEnvelope.cs
@@ -20,7 +20,7 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-namespace EventFlow.ReadStores.V2
+namespace EventFlow.ReadStores
 {
     public class ReadModelEnvelope<TReadModel>
         where TReadModel : class, IReadModel, new()
