+++ Source/EventFlow/ReadStores/SingleAggregateReadStoreManager.cs
@@ -20,7 +20,9 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.Aggregates;
@@ -28,43 +30,46 @@ using EventFlow.Logs;
 
 namespace EventFlow.ReadStores
 {
-    public abstract class SingleAggregateReadModelStore<TReadModel> :
-        ReadModelStore<TReadModel>
-        where TReadModel : IReadModel
+    public class SingleAggregateReadStoreManager<TReadModelStore, TReadModel> : ReadStoreManager<TReadModelStore, TReadModel>
+        where TReadModelStore : IReadModelStore<TReadModel>
+        where TReadModel : class, IReadModel, new()
     {
-        protected SingleAggregateReadModelStore(
+        public SingleAggregateReadStoreManager(
             ILog log,
+            TReadModelStore readModelStore,
             IReadModelDomainEventApplier readModelDomainEventApplier)
-            : base(log, readModelDomainEventApplier)
+            : base(log, readModelStore, readModelDomainEventApplier)
         {
         }
 
-        public override Task ApplyDomainEventsAsync(
-            IReadOnlyCollection<IDomainEvent> domainEvents,
-            CancellationToken cancellationToken)
+        protected override IReadOnlyCollection<ReadModelUpdate> BuildReadModelUpdates(
+            IReadOnlyCollection<IDomainEvent> domainEvents)
         {
-            throw new System.NotImplementedException();
+            var readModelIds = domainEvents
+                .Select(e => e.GetIdentity().Value)
+                .Distinct()
+                .ToList();
+            if (readModelIds.Count != 1)
+            {
+                throw new ArgumentException("Only domain events from the same aggregate is allowed");
+            }
+
+            return new[] {new ReadModelUpdate(readModelIds.Single(), domainEvents)};
         }
 
-        public override Task ApplyDomainEventsAsync<TReadModelToPopulate>(
-            IReadOnlyCollection<IDomainEvent> domainEvents,
-            CancellationToken cancellationToken)
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public override Task PurgeAsync<TReadModelToPurge>(CancellationToken cancellationToken)
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public override Task PopulateReadModelAsync<TReadModelToPopulate>(
-            string id,
-            IReadOnlyCollection<IDomainEvent> domainEvents,
+        protected override async Task<ReadModelEnvelope<TReadModel>> UpdateAsync(
             IReadModelContext readModelContext,
+            IReadOnlyCollection<IDomainEvent> domainEvents,
+            ReadModelEnvelope<TReadModel> readModelEnvelope,
             CancellationToken cancellationToken)
         {
-            throw new System.NotImplementedException();
+            var readModel = readModelEnvelope.ReadModel ?? new TReadModel();
+
+            await ReadModelDomainEventApplier.UpdateReadModelAsync(readModel, domainEvents, readModelContext, cancellationToken).ConfigureAwait(false);
+
+            var readModelVersion = domainEvents.Max(e => e.AggregateSequenceNumber);
+
+            return ReadModelEnvelope<TReadModel>.With(readModel, readModelVersion);
         }
     }
 }
